Design Specification for Supply Chain Management
Value Added Trading (VAT)
The core requirement for any successful supply chain management would be to have the right products in the right quantities at the right place at the right moment at minimum cost, with utmost customer satisfaction, inventory management, and flexibility. A product supplier's flexibility is its ability to respond to changes in demand. A company that can receive orders from Customers, process the order and receive components from its Suppliers, assemble these, fill the orders, and ship the finished products to customers in one single day. This company would have total flexibility. It would be able to prevail under any unforeseen events on a day-to-day basis and could easily attain a hundred percent customer satisfaction without any inventory.

The central aim of this supply chain management research study is to have organizations have the right products in the right quantities (at the right place) at the right moment at minimal cost, which is translated into the interrelated issues of customer satisfaction, inventory management, and flexibility. Customer satisfaction is to a high degree dependent on the flexibility of the supply chain, i.e. its ability to respond to changes in demand. Flexibility is often imperfect because of long lead times, uncertainties, and unforeseen events. To counterbalance this lack of flexibility companies will keep inventories at various levels of the supply chain. Balancing the costs of imperfect customer satisfaction and holding inventory is a classic issue of supply chain management. With the customer-centric distribution chain today, the role that a distributor plays includes enabling direct ordering and tracking of orders over the Internet.